add an optional time_shift variable in the input JSON file
This allows for the overlay date time to be shifted, to account for
possible mismatch in the time offset with respect to the background image.
Note, had to remove the line in satpy_overlay_plots that recomputes the datetime
based on the overlay data - this was already done in the parent function, so
this was redundant, and would undo the time shift.